refactor: move BranchAction interface into a single file
